requirements.txt
fastapi
uvicorn
httpx
google-generativeai
python-dotenv
rich
pyfiglet
